Delete User Goal

+++ server/controllers/goalsController.js
@@ -70,6 +70,19 @@ const deleteGoal = asyncHandler(async (req, res) => {
     throw new Error('Goal not found')
   }
 
+  const user = await User.findById(req.user.id);
+
+  // check for user
+  if(!user){
+    res.status(401)
+    throw new Error('User not found')
+  }
+  // Make sure the logged in user matches the goal user
+  if(goal.user.toString() !== user.id) {
+    res.status(401)
+    throw new Error ('User not authorized');
+  }
+
   await goal.deleteOne(user);
 
   res.status(200).json({ id: req.params.id })
